impl: Implement a v2 of the freelist allocator
This version of the freelist allocator is implemented with
an alternate layout that keeps the size of the type consistent
with the size of a pointer.

Additionally, each object is contained in its own malloc() allocation,
which is lazily reused. This might be a bit worse off, which is why
both v1 and v2 are currently in the repository.